PD#137490: Security patch from bulletin-2016-10.
Security Vulnerability - CVE-2012-6702 and CVE-2016-5300

Applies the patch from the bug. The change to CMakeLists.txt is
not applicable as Android does not use that to build expat. It
is not needed as it only applies to building for Windows.

Bug: 29149404
Change-Id: I557d2d5e5f338cd141db7f09fb2648f1be4cc8bc
(cherry picked from commit 3c2f09e63ae65adffdfe07164f872e454e338789)